Create tkinter - python
in this project, i combined pygame with tkinter, first i want the player to input the the correct password and name, after that, you click the "đăng nhập" button. This will lead to another window and that window show you a another button that lead to the pygame window.